+++ app/templates/app/app.js
@@ -1,0 +1,3 @@
+(function(logger) {
+  logger.log('Hello, <%= projectName %>');
+}(console));