+++ app/templates/grunt/tasks/clean.js
@@ -1,0 +1,18 @@
+/**
+ * Clean task for temporary, dist, doc and test results folders.
+ *
+ * @param {Object} grunt Reference to the current Grunt process.
+ */
+module.exports = function(grunt) {
+  grunt.loadNpmTasks('grunt-contrib-clean');
+
+  grunt.config.set('clean', {
+    options: {
+      force: true
+    },
+    test: ['<%%= source.test %>/results'],
+    tmp: ['<%%= source.tmp %>'],
+    dist: ['<%%= source.dist %>'],
+    doc: ['<%%= source.doc %>']
+  });
+};