+++ app/templates/grunt/tasks/concurrent.js
@@ -1,0 +1,17 @@
+/**
+ * Setup concurrent tasks.
+ *
+ * @param {Object} grunt Reference to the current Grunt process.
+ */
+module.exports = function(grunt) {
+  grunt.loadNpmTasks('grunt-concurrent');
+
+  grunt.config.set('concurrent', {
+    options: {
+      logConcurrentOutput: true
+    },
+    default: [
+      'watch'
+    ]
+  });
+};