+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # 1.0.0
 
 - Initial version of generator-build-tools
+- Added grunt tasks
+- Added gulp tasks
+- Added LESS, SASS, Stylus support